calculator

+++ 230921/230921/lecture2.cpp
@@ -4,7 +4,7 @@ double sum(double a, double b);
 double sub(double a, double b);
 double mul(double a, double b);
 double div(double a, double b);
-void calculator(double selection);
+void calculator(int selection);
 
 int main()
 {
@@ -15,7 +15,7 @@ int main()
 	return 0;
 }
 
-void calculator(double selection) {
+void calculator(int selection) {
 	double num1, num2;
 	double result;
